+++ example/src/App.tsx
@@ -41,6 +41,8 @@ import PayPalScreen from './screens/PayPalScreen';
 import AffirmScreen from './screens/AffirmScreen';
 import CollectBankAccountScreen from './screens/CollectBankAccountScreen';
 import CashAppScreen from './screens/CashAppScreen';
+import PaymentSheetDeferredIntentScreen from './screens/PaymentSheetDeferredIntentScreen';
+import PaymentSheetDeferredIntentMultiStepScreen from './screens/PaymentSheetDeferredIntentMultiStepScreen';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
@@ -83,6 +85,8 @@ export type RootStackParamList = {
   CashAppScreen: undefined;
   AffirmScreen: undefined;
   CollectBankAccountScreen: undefined;
+  PaymentSheetDeferredIntentScreen: undefined;
+  PaymentSheetDeferredIntentMultiStepScreen: undefined;
 };
 
 declare global {
@@ -148,6 +152,14 @@ export default function App() {
           <Stack.Screen
             name="PaymentSheetWithSetupIntent"
             component={PaymentSheetWithSetupIntent}
+          />
+          <Stack.Screen
+            name="PaymentSheetDeferredIntentScreen"
+            component={PaymentSheetDeferredIntentScreen}
+          />
+          <Stack.Screen
+            name="PaymentSheetDeferredIntentMultiStepScreen"
+            component={PaymentSheetDeferredIntentMultiStepScreen}
           />
           <Stack.Screen
             name="PaymentsUICustomScreen"