+++ ios/StripeSdk.swift
@@ -2,6 +2,7 @@ import PassKit
 import Stripe
 import StripePaymentSheet
 import StripeFinancialConnections
+import Foundation
 
 @objc(StripeSdk)
 class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdaptivePresentationControllerDelegate {
@@ -12,16 +13,11 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
 
     internal var paymentSheet: PaymentSheet?
     internal var paymentSheetFlowController: PaymentSheet.FlowController?
-
+    var paymentSheetIntentCreationCallback: ((Result<String, Error>) -> Void)?
+    
     var urlScheme: String? = nil
     
     var confirmPaymentResolver: RCTPromiseResolveBlock? = nil
-
-    var applePayCompletionCallback: STPIntentClientSecretCompletionBlock? = nil
-    var deprecatedApplePayRequestResolver: RCTPromiseResolveBlock? = nil
-    var deprecatedApplePayRequestRejecter: RCTPromiseRejectBlock? = nil
-    var deprecatedApplePayCompletionRejecter: RCTPromiseRejectBlock? = nil
-    var deprecatedConfirmApplePayPaymentResolver: RCTPromiseResolveBlock? = nil
     
     var confirmApplePayResolver: RCTPromiseResolveBlock? = nil
     var confirmApplePayPaymentClientSecret: String? = nil
@@ -52,16 +48,16 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
     var applePayShippingAddressErrors: [Error]? = nil
     var applePayCouponCodeErrors: [Error]? = nil
     
-    var hasLegacyApplePayListeners = false
+    var hasEventListeners = false
     override func startObserving() {
-        hasLegacyApplePayListeners = true
+        hasEventListeners = true
     }
     override func stopObserving() {
-        hasLegacyApplePayListeners = false
+        hasEventListeners = false
     }
     
     override func supportedEvents() -> [String]! {
-        return ["onDidSetShippingMethod", "onDidSetShippingContact", "onDidSetCouponCode"]
+        return ["onOrderTrackingCallback", "onConfirmHandlerCallback"]
     }
 
     @objc override static func requiresMainQueueSetup() -> Bool {
@@ -109,8 +105,6 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
     @objc(initPaymentSheet:resolver:rejecter:)
     func initPaymentSheet(params: NSDictionary, resolver resolve: @escaping RCTPromiseResolveBlock,
                           rejecter reject: @escaping RCTPromiseRejectBlock) -> Void  {
-        self.paymentSheetFlowController = nil
-        
         let (error, configuration) = buildPaymentSheetConfiguration(params: params)
         guard let configuration = configuration else {
             resolve(error)
@@ -120,16 +114,29 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
         preparePaymentSheetInstance(params: params, configuration: configuration, resolve: resolve)
     }
     
-    @objc(initPaymentSheetWithOrderTracking:callback:resolver:rejecter:)
-    func initPaymentSheetWithOrderTracking(params: NSDictionary, callback orderTrackingCallback: @escaping RCTResponseSenderBlock, resolver resolve: @escaping RCTPromiseResolveBlock,
+    @objc(intentCreationCallback:resolver:rejecter:)
+    func intentCreationCallback(result: NSDictionary, resolver resolve: @escaping RCTPromiseResolveBlock,
                           rejecter reject: @escaping RCTPromiseRejectBlock) -> Void  {
-        let (error, configuration) = buildPaymentSheetConfiguration(params: params, orderTrackingCallback: orderTrackingCallback)
-        guard let configuration = configuration else {
-            resolve(error)
+        guard let paymentSheetIntentCreationCallback = self.paymentSheetIntentCreationCallback else {
+            resolve(Errors.createError(ErrorType.Failed, "No intent creation callback was set"))
             return
         }
-        
-        preparePaymentSheetInstance(params: params, configuration: configuration, resolve: resolve)
+        if let clientSecret = result["clientSecret"] as? String {
+            paymentSheetIntentCreationCallback(.success(clientSecret))
+        } else {
+            class ConfirmationError: Error, LocalizedError {
+                private var errorMessage: String
+                init(errorMessage: String) {
+                    self.errorMessage = errorMessage
+                }
+                public var errorDescription: String? {
+                    return errorMessage
+                }
+            }
+            let errorParams = result["error"] as? NSDictionary
+            let error = ConfirmationError.init(errorMessage: errorParams?["localizedMessage"] as? String ?? "An unknown error occurred.")
+            paymentSheetIntentCreationCallback(.failure(error))
+        }
     }
 
     @objc(confirmPaymentSheetPayment:rejecter:)
@@ -342,54 +349,11 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
         self.shippingContactUpdateCompletion = nil
         resolve([])
     }
-    
-    @objc(updateApplePaySummaryItems:errorAddressFields:resolver:rejecter:)
-    func updateApplePaySummaryItems(summaryItems: NSArray, errorAddressFields: [NSDictionary], resolver resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-        if (shippingMethodUpdateCompletion == nil && shippingContactUpdateCompletion == nil) {
-            resolve(Errors.createError(ErrorType.Failed, "You can use this method only after either onDidSetShippingMethod or onDidSetShippingContact events emitted"))
-            return
-        }
-
-        var paymentSummaryItems : [PKPaymentSummaryItem] = []
-        do {
-            paymentSummaryItems = try ApplePayUtils.buildPaymentSummaryItems(items: summaryItems as? [[String : Any]])
-        } catch {
-            resolve(Errors.createError(ErrorType.Failed, error.localizedDescription))
-            return
-        }
-
-        var shippingAddressErrors: [Error] = []
-        for item in errorAddressFields {
-            let field = item["field"] as! String
-            let message = item["message"] as? String ?? field + " error"
-            shippingAddressErrors.append(PKPaymentRequest.paymentShippingAddressInvalidError(withKey: field, localizedDescription: message))
-        }
-
-        shippingMethodUpdateCompletion?(PKPaymentRequestShippingMethodUpdate.init(paymentSummaryItems: paymentSummaryItems))
-        shippingContactUpdateCompletion?(PKPaymentRequestShippingContactUpdate.init(errors: shippingAddressErrors, paymentSummaryItems: paymentSummaryItems, shippingMethods: []))
-        self.shippingMethodUpdateCompletion = nil
-        self.shippingContactUpdateCompletion = nil
-        resolve([])
-    }
 
     @objc(openApplePaySetup:rejecter:)
     func openApplePaySetup(resolver resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) -> Void {
         PKPassLibrary.init().openPaymentSetup()
         resolve([])
-    }
-
-    @objc(confirmApplePayPayment:resolver:rejecter:)
-    func confirmApplePayPayment(clientSecret: String, resolver resolve: @escaping RCTPromiseResolveBlock, rejecter reject: @escaping RCTPromiseRejectBlock) {
-        self.deprecatedApplePayCompletionRejecter = reject
-        self.deprecatedConfirmApplePayPaymentResolver = resolve
-        self.applePayCompletionCallback?(clientSecret, nil)
-    }
-
-    @objc(isApplePaySupported:rejecter:)
-    func isApplePaySupported(resolver resolve: @escaping RCTPromiseResolveBlock,
-                             rejecter reject: @escaping RCTPromiseRejectBlock) {
-        let isSupported = StripeAPI.deviceSupportsApplePay()
-        resolve(isSupported)
     }
 
     @objc(handleURLCallback:resolver:rejecter:)
@@ -407,28 +371,6 @@ class StripeSdk: RCTEventEmitter, STPBankSelectionViewControllerDelegate, UIAdap
           resolve(stripeHandled)
         }
       }
-    }
-
-    @objc(presentApplePay:resolver:rejecter:)
-    func presentApplePay(params: NSDictionary,
-                         resolver resolve: @escaping RCTPromiseResolveBlock,
-                         rejecter reject: @escaping RCTPromiseRejectBlock) {
-        self.deprecatedApplePayRequestResolver = resolve
-        self.deprecatedApplePayRequestRejecter = reject
-        
-        let (error, paymentRequest) = ApplePayUtils.createPaymentRequest(merchantIdentifier: merchantIdentifier, params: params)
-        guard let paymentRequest = paymentRequest else {
-            resolve(error)
-            return
-        }
-        
-        if let applePayContext = STPApplePayContext(paymentRequest: paymentRequest, delegate: self) {
-            DispatchQueue.main.async {
-                applePayContext.presentApplePay(completion: nil)
-            }
-        } else {
-            resolve(Errors.createError(ErrorType.Failed, "Payment not completed"))
-        }
     }
     
     @objc(isPlatformPaySupported:resolver:rejecter:)