+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 ## Unreleased
 
+### Breaking changes
+
+- The Apple Pay and Google Pay APIs, which are deprecated and were replaced with the [Platform Pay API](https://github.com/stripe/stripe-react-native/blob/master/docs/Platform-Pay-Migration.md) last year, have been removed.
+
+## Features
+
+- You can now collect payment details before creating a `PaymentIntent` or `SetupIntent`. See [our docs](https://stripe.com/docs/payments/accept-a-payment-deferred?platform=react-native) for more info. This integration also allows you to [confirm the Intent on the server]
+(https://stripe.com/docs/payments/finalize-payments-on-the-server?platform=react-native).
+
 ## Fixes
 
-- Only call `handleURLCallback` on iOS. [#1423](https://github.com/stripe/stripe-react-native/pull/1423)
+- Fixes `handleURLCallback` to only take action on iOS, no-op on Android. [#1423](https://github.com/stripe/stripe-react-native/pull/1423)
 
 ## 0.28.0 - 2023-06-16